# Detailed and CSV Output + 