--- a/src/main.c
+++ b/src/main.c
@@ -1,23 +1,35 @@
 #include <ecs_benchmark.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-#define WARMUP_INTERVALS (5)
+// How many times a benchmark is run before recording a measurement.
 #define MEASURE_INTERVAL (50)
-#define MEASURE_TIME (0.5)
+// The minimum amount of time a benchmark must have run for.
+#define MEASURE_TIME (1.0)
+// The number of times to run the benchmark before starting to record.
+#define WARMUP_INTERVAL (100)
 
 #define PRETTY_TIME_FMT
 
+typedef enum {
+    BenchTable,
+    BenchTableDetailed,
+    BenchCSV,
+} bench_result_format_t;
+
 typedef struct bench_t {
     // Benchmark state
-    // The number of times a benchmark is run before taking a measurement.
+    // The number of times a benchmark is run before recording a measurement.
     uint32_t interval;
-    // The total number of intervals that have been run.
-    uint32_t intervals;
-    // The time at which the benchmark was started.
-    ecs_time_t t;
-    // The time it took to run the benchmark.
-    double dt;
+    // The total number of samples that have been run.
+    uint32_t samples;
+    // The amount of time the benchmark was running.
+    ecs_time_t total;
+    // The time at which the benchmark started.
+    ecs_time_t start;
+    // The time at which the benchmark finished.
+    ecs_time_t stop;
 
     // Benchmark settings
     // The label of the benchmark.
@@ -46,27 +58,12 @@ typedef struct bench_t {
 #define COLOR(v)\
     ((v < FRAC_MILLION) ? ECS_GREEN : (v < FRAC_THOUSAND) ? ECS_YELLOW : ECS_RED)
 
+bench_result_format_t g_format = BenchTable;
+bool g_detailed = false;
+
 bool flip_coin(void) {
     int r = rand();
     return r >= ((float)RAND_MAX / 2.0f);
-}
-
-const char* spacing(float v) {
-    if (v > 1000000) {
-        return "";
-    } else if (v > 100000) {
-        return " ";
-    } else if (v > 10000) {
-        return "  ";
-    } else if (v > 1000) {
-        return "   ";
-    } else if (v > 100) {
-        return "    ";
-    } else if (v > 10) {
-        return "     ";
-    } else {
-        return "      ";
-    }
 }
 
 char* bench_vasprintf(
@@ -109,63 +106,116 @@ char* bench_asprintf(
     return result;
 }
 
-#ifdef PRETTY_TIME_FMT
 void header_print(void) {
-    printf("| Benchmark                             | Measurement  |\n");
-    printf("|---------------------------------------|--------------|\n");
+    switch (g_format) {
+    case BenchTable:
+        if (g_detailed) {
+            printf("| Benchmark                             | Measurement  | Total (ns)   | Count          | Samples      |\n");
+            printf("|---------------------------------------|--------------|--------------|----------------|--------------|\n");
+        } else {
+            printf("| Benchmark                             | Measurement  |\n");
+            printf("|---------------------------------------|--------------|\n");
+        }
+        break;
+    case BenchCSV:
+        if (g_detailed) {
+            printf("Benchmark,Measurement (ns),Total (ns),Count,Samples\n");
+        } else {
+            printf("Benchmark,Measurement (ns)\n");
+        }
+        break;
+    }
 }
-
-void bench_print(const char *label, float v) {
-    printf("| %s %*s | %s%.2f%s%s%s |\n", 
-        label, (int)(36 - strlen(label)), "", COLOR(v), NUM(v), SYM(v), spacing(NUM(v)), ECS_NORMAL);
-}
-#else
-void header_print(void) {
-    printf("Benchmark                              Measurement\n");
-}
-
-void bench_print(const char *label, float v) {
-    printf("%s %*s  %s%.2f%s%s\n", 
-        label, (int)(36 - strlen(label)), "", COLOR(v), v * BILLION, spacing(v * BILLION), ECS_NORMAL);
-}
-#endif
 
 bench_t bench_begin(const char *lbl, int32_t count) {
     bench_t b = {0};
     b.lbl = lbl;
-    b.interval = MEASURE_INTERVAL;
-    b.intervals = 0;
+    b.interval = WARMUP_INTERVAL;
+    b.samples = 0;
     b.count = count;
     return b;
 }
 
-double time_measure(
-    ecs_time_t *start)
+ecs_time_t ecs_time_add(
+    ecs_time_t t1,
+    ecs_time_t t2)
 {
-    ecs_time_t stop;
-    ecs_os_get_time(&stop);
-    return ecs_time_to_double(ecs_time_sub(stop, *start));
+    ecs_time_t result;
+    result.nanosec = t1.nanosec + t2.nanosec;
+    result.sec = t1.sec + t2.sec;
+
+    if (result.nanosec >= 1000000000) {
+        uint32_t s = result.nanosec / 1000000000;
+        result.nanosec -= s * 1000000000;
+        result.sec = s;
+    }
+
+    return result;
+}
+
+uint64_t ecs_time_to_nanos(ecs_time_t t) {
+    return t.sec * 1000000000 + t.nanosec;
 }
 
 bool bench_next(bench_t *b) {
     if (!--b->interval) {
-        b->intervals ++;
-        if (b->intervals > WARMUP_INTERVALS) {
-            double dt = time_measure(&b->t);
-            if (dt > MEASURE_TIME) {
-                b->dt = dt;
+        // Record the stop time as soon as possible (to reduce overhead)
+        ecs_os_get_time(&b->stop);
+
+        // `samples == 0` means we are doing warmup
+        // When `samples > 0` record a measurement
+        if (b->samples++) {
+            // Accumulate the delta time
+            b->total = ecs_time_add(b->total, ecs_time_sub(b->stop, b->start));
+            // TODO: Record each individual sample so we can calculate percentiles, variance, etc
+
+            // Check if we have run long enough
+            if (ecs_time_to_nanos(b->total) > (uint64_t)(MEASURE_TIME * BILLION)) {
                 return false;
             }
-        } else if (b->intervals == WARMUP_INTERVALS) {
-            ecs_os_get_time(&b->t);
         }
         b->interval = MEASURE_INTERVAL;
+
+        // Record the start time as late as possible (to reduce overhead)
+        ecs_os_get_time(&b->start);
     }
     return true;
 }
 
 void bench_end(bench_t *b) {
-    bench_print(b->lbl, b->dt / ((b->intervals - WARMUP_INTERVALS) * MEASURE_INTERVAL * b->count));
+    const char* label = b->lbl;
+    // Note: -1 because we don't count the warmup
+    uint64_t count = ((uint64_t)b->samples - 1) * (uint64_t)MEASURE_INTERVAL * (uint64_t)b->count;
+    uint64_t total_nanos = ecs_time_to_nanos(b->total);
+
+    switch (g_format) {
+    case BenchTable:
+        // Pretty print the measurement
+        double total_secs = ecs_time_to_double(b->total);
+        double v = total_secs / count;
+        if (g_detailed) {
+            printf("| %-37s | %s%10.3f%s%s | %12" PRIu64 " | %14" PRIu64 " | %12" PRIu32 " |\n", 
+                label, 
+                COLOR(v), NUM(v), SYM(v), ECS_NORMAL,
+                total_nanos,
+                count,
+                b->samples - 1);
+        } else {
+            printf("| %-37s | %s%10.3f%s%s |\n", 
+                label, 
+                COLOR(v), NUM(v), SYM(v), ECS_NORMAL);
+        }
+        break;
+    case BenchCSV:
+        if (g_detailed) {
+            double v = (double)total_nanos / count;
+            printf("%s,%f,%" PRIu64 ",%" PRIu64 ",%" PRIu32 "\n", label, v, total_nanos, count, b->samples - 1);
+        } else {
+            double v = (double)total_nanos / count;
+            printf("%s,%f\n", label, v);
+        }
+        break;
+    }
 }
 
 /* -- benchmark code -- */
@@ -197,9 +247,14 @@ ecs_entity_t* create_ids(ecs_world_t *world, int32_t count, ecs_size_t size, boo
 
 void baseline(void) {
     bench_t b = bench_begin("baseline", 1);
+    uint32_t result = 0;
     do {
+        result++;
     } while (bench_next(&b));
     bench_end(&b);
+
+    printf("result = %u\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b", 
+        result);
 }
 
 void world_mini_fini(void) {
@@ -1816,6 +1871,14 @@ void query_count(const char *label, int32_t table_count) {
 
 int main(int argc, char *argv[]) {
     ecs_os_set_api_defaults(); // Required for timers to work
+
+    for (int i = 0; i < argc; i++) {
+        if (!strcmp(argv[i], "--csv")) {
+            g_format = BenchCSV;
+        } else if (!strcmp(argv[i], "--detailed")) {
+            g_detailed = true;
+        }
+    }
 
     header_print();