@SanderMertens 2024-01-07
Hm, the example outputs don't look right:

```c
| get_empty_entity                      |      0.002ns |
| get_id_not_found                      |      0.002ns |
| get_id                                |      0.002ns |
| get_16_ids                            |      0.000ns |
| get_pair                              |      0.002ns |
| get_pair_16_targets                   |      0.000ns |
```
These operations definitely take longer than 0.002 nanoseconds 😅