@ZeroErrors 2024-01-08
hmm, I'm getting this data pretty consistently but I'll do some more digging.

Edit: I found the cause of the large disparity, but will continue doing more testing.

**Before:**
Running on master I get values like this.
```md
| Benchmark                             | Measurement  |
|---------------------------------------|--------------|
| get_empty_entity                      | 0.12ns       |
| get_id_not_found                      | 0.14ns       |
| get_id                                | 7.27ns       |
| get_16_ids                            | 0.13ns       |
| get_pair                              | 0.32ns       |
| get_pair_16_targets                   | 1.71ns       |
```
Tweaked it to include wall time which is a timer measured from `bench_begin` to `bench_end` to get a bit more info.

```md
| Benchmark                             | Measurement  | dt (ns)        | Count          | Intervals      | Bench Count    | Wall Time      |
|---------------------------------------|--------------|----------------|----------------|----------------|----------------|----------------|
| get_empty_entity                      | 0.51ns       | 500000100.0000 |      971840192 |       12732540 |           1000 |      500013100 
| get_id_not_found                      | 0.41ns       | 500000100.0000 |     1224271008 |       13081186 |           1000 |      500000200 
| get_id                                | 0.12ns       | 500000100.0000 |     4190255600 |       12968707 |           1000 |      500000500 
| get_16_ids                            | 0.80ns       | 500000100.0000 |      625412864 |       12971583 |          16000 |      500000400 
| get_pair                              | 0.14ns       | 500000100.0000 |     3471257488 |       12696629 |           1000 |      500000300 
| get_pair_16_targets                   | 0.16ns       | 500000100.0000 |     3213355008 |       12899656 |          16000 |      500000400 
```

**After:**
Running with the changes in this PR I get something like the following.
This also is modified to add an additional column, wall time.
```md
| Benchmark                             | Measurement  | Total (ns)   | Count          | Samples      | Interval     | Wall Time (ns) |
|---------------------------------------|--------------|--------------|----------------|--------------|--------------|----------------|
| get_empty_entity                      |      0.001ns |    506450800 |   450000000000 |            9 |     50000000 |      506452500 |
| get_id_not_found                      |      0.001ns |    504620300 |   450000000000 |            9 |     50000000 |      504622200 |
| get_id                                |      0.002ns |    550030300 |   350000000000 |            7 |     50000000 |      550031700 |
| get_16_ids                            |      0.000ns |    551502200 |  5600000000000 |            7 |     50000000 |      551503600 |
| get_pair                              |      0.001ns |    557665700 |   450000000000 |            9 |     50000000 |      557667600 |
| get_pair_16_targets                   |      0.000ns |    560586900 |  7200000000000 |            9 |     50000000 |      560588600 |
```

So looking at `get_id` since it has the largest variance.
Before:
```c
bench_count = 1000
measurement = 0.12ns
total = 500000100.0000ns
count = 4190255600
intervals = 12968707
wall_time = 500000500ns

count = intervals * MEASURE_INTERVAL * bench_count
count = 12968707 * 50 * 1000
count = 648435350000 != 4190255600 // WRONG: !!!!! >32-bit

overhead = wall_time - total
overhead = 500000500ns - 500000100ns
overhead = 400ns

measurement_wall_time = wall_time  / count
measurement_wall_time_fixed = 500000500ns / 648435350000 // Using the above calculated value
measurement_wall_time_fixed = 0.0007710876651ns
measurement_wall_time_wrong = 500000500ns / 4190255600   // Using the wrong value
measurement_wall_time_wrong = 0.119324582491ns == 0.12ns

measurement = total / count
measurement_fixed = 500000100.0000ns / 648435350000 // Using the above calculated value
measurement_fixed = 0.000771087048231ns 
measurement_wrong = 500000100.0000ns / 4190255600  // Using the wrong value
measurement_wrong = 0.119324487031ns == 0.12ns
```
This code seems to be the cause.
```c++
= b->dt / ((b->intervals - WARMUP_INTERVALS) * MEASURE_INTERVAL * b->count)

= double / ((uint32_t - int) * int * uint32_t)
= double / (uint32_t * int * uint32_t)
= double / uint32_t
= double
```

If I change it to cast the intervals to `uint64_t` eg. `b->dt / (((uint64_t)b->intervals - WARMUP_INTERVALS) * MEASURE_INTERVAL * b->count)` I get a measurement around `0.00076563ns` instead, which is actually lower than what my changes show.. hmm...


After:
```c
bench_count = 1000
measurement = 0.002ns
total = 546608600ns
count = 350000000000
wall_time = 546610100ns

count = samples * sample_interval * bench_count
count = 7 * 50000000 * 1000
count = 350000000000 // Expected

overhead = wall_time - total
overhead = 546610100ns - 546608600ns
overhead = 1500ns

measurement_wall_time = wall_time  / count
measurement_wall_time = 546610100ns / 350000000000
measurement_wall_time = 0.00156174314286ns == 0.002ns // Expected

measurement = total / count
measurement = 546608600ns / 350000000000
measurement = 0.00156173885714ns == 0.002ns // Expected
```