@SanderMertens 2024-01-08
Any measurement that's below one nanosecond is _very_ suspect. Try to cross-reference the results with the README, they should be roughly in the same ballpark.